#49 - moving filter to be covered by api version
