+++ trade-module/src/main/java/io/edpn/backend/trade/domain/filter/v1/LocateCommodityFilter.java
@@ -1,4 +1,4 @@
-package io.edpn.backend.trade.domain.filter;
+package io.edpn.backend.trade.domain.filter.v1;
 
 import io.edpn.backend.trade.domain.model.LandingPadSize;
 import lombok.AllArgsConstructor;
