+++ trade-module/src/main/java/io/edpn/backend/trade/domain/usecase/LocateCommodityUseCase.java
@@ -1,7 +1,7 @@
 package io.edpn.backend.trade.domain.usecase;
 
 import io.edpn.backend.trade.domain.model.LocateCommodity;
-import io.edpn.backend.trade.domain.filter.LocateCommodityFilter;
+import io.edpn.backend.trade.domain.filter.v1.LocateCommodityFilter;
 
 import java.util.List;
